move extract_point_from_geometry function as a method in its own class
